pyproject.toml
[tool.poetry]
name = "data-synthesis-example"
version = "0.1.0"
description = ""
authors = ["majimearun <arunachala0619@gmail.com>"]
readme = "README.md"
packages = [{include = "data_synthesis_example"}]

[tool.poetry.dependencies]
python = "^3.10"
pandasai = "^0.6.2"
faker = "^18.11.1"
google-generativeai = "^0.1.0"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
tabulate = "^0.9.0"
pyarrow = "^12.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
